fix(quorum): address CodeRabbit review on checkpoint-quorum (#1264)
- threshold: tighten the public API of checkpoint_quorum_message /
  verify_checkpoint_quorum / cosign_checkpoint from usize to u32 so the value
  bound into the fixed 4-byte BLAKE3 field cannot silently truncate (CR Major).
  The big-endian framing is unchanged, so the golden vector and JS verifier are
  byte-identical (confirmed: git diff on the vector is empty). The golden-vector
  test now validates the u32 bound via u32::try_from.
- JS verifier: wrap cosignature field parsing in try/catch so a malformed entry
  is skipped instead of aborting the case, matching Rust verify_checkpoint_quorum's
  fail-closed `let (Ok(..), ..) = parse_fr else { continue; }` (CR Major).

Skipped (verified against repo conventions):
- "canonical JSON / add serde_jcs for the fixture": every existing fixture
  (vectors.json 2403 lines, redaction_vectors.json, parser_vectors.json) is
  pretty-printed; RFC-8785/JCS governs hashed/signed protocol JSON, not parsed
  test fixtures whose bytes are never hashed. No new (possibly-abandoned) dep.
- "fold into vectors.json as the single source": the repo uses per-concern
  fixture files (parser_vectors.json, redaction_vectors.json); a sibling
  checkpoint_quorum_vectors.json matches that established pattern.

Verified: cargo test -p olympus-desktop --lib checkpoint (14 pass);
node verifiers/javascript/test_checkpoint_quorum.js (8 pass); vector byte-identical.

+++ src-tauri/examples/gen_checkpoint_quorum_vectors.rs
@@ -42,7 +42,7 @@ fn case(
     name: &str,
     doc: &str,
     root: &Fr,
-    threshold: usize,
+    threshold: u32,
     signers: &[QuorumSigner],
     cosignatures: &[CollectedSignature],
 ) -> Value {
@@ -79,7 +79,7 @@ fn main() {
     let set2 = vec![s1.clone(), s2.clone()];
     let set1 = vec![s1.clone()];
 
-    let cosign = |k: &[u8; 32], root: &Fr, t: usize, set: &[QuorumSigner]| {
+    let cosign = |k: &[u8; 32], root: &Fr, t: u32, set: &[QuorumSigner]| {
         cosign_checkpoint(k, root, t, set).expect("cosign")
     };
